Add files via upload
Life Years Lost (LYL) Metric added to clean wonder.  LYL is like Deaths, it is another metric to measure the impact of the drug crisis.  
Note: 2 rows (subgroups) can have the same crude rate (Deaths/Population) but can have different LYL, if the age of the row is different.  eg. if the age of death is 30, there will be more life years lost than if the age of death is 50.